Feat: implement find member whole summary function by admin
