+++ src/main/java/kr/codeback/controller/AdminController.java
@@ -1,78 +1,41 @@
 package kr.codeback.controller;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CookieValue;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import java.util.ArrayList;
+import java.util.Map;
 
-import kr.codeback.common.MessageResponseDTO;
-import kr.codeback.model.constant.SuccessMessage;
-import kr.codeback.model.dto.request.AuthorityRequestDTO;
-import kr.codeback.model.dto.response.AuthorityResponseDTO;
-import kr.codeback.model.dto.response.MembersWithPageResponseDTO;
-import kr.codeback.model.entity.Authority;
-import kr.codeback.model.entity.Member;
-import kr.codeback.service.interfaces.AuthorityService;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+import kr.codeback.model.dto.response.MemberSummaryResponseDTO;
 import kr.codeback.service.interfaces.MemberService;
-import kr.codeback.util.JwtUtil;
 import lombok.RequiredArgsConstructor;
 
-@RestController
-@RequestMapping("/api/admin")
+@Controller
 @RequiredArgsConstructor
 public class AdminController {
 
-	private final JwtUtil jwtUtil;
 	private final MemberService memberService;
-	private final AuthorityService authorityService;
 
-	@GetMapping("/members")
-	public ResponseEntity<MembersWithPageResponseDTO> findAllMembers(
-		@RequestParam(required = false, defaultValue = "0", value = "pageNum") int pageNum,
-		@RequestParam(required = false, defaultValue = "10", value = "pageSize") int pageSize,
-		@CookieValue(value = "access_token") String jwtToken) {
+	@GetMapping("/admin/members")
+	public String moveMembers(Model model) {
 
-		String email = jwtUtil.extractEmail(jwtToken);
-		memberService.validateAdminMemberByEmail(email);
+		model.addAttribute("initialMembers", new ArrayList<>());
 
-		MembersWithPageResponseDTO membersWithPageResponseDTO = memberService.findAllUnderAdmin(pageNum,
-			pageSize);
-
-		return ResponseEntity.status(HttpStatus.OK).body(membersWithPageResponseDTO);
+		return "/view/admin/members";
 	}
 
-	@DeleteMapping("/member/{email}")
-	public ResponseEntity<MessageResponseDTO> deleteMember(@PathVariable(name = "email") String deleteEmail,
-		@CookieValue(value = "access_token") String jwtToken) {
+	@GetMapping("/admin/summary")
+	public String moveSummary(Model model) {
 
-		String email = jwtUtil.extractEmail(jwtToken);
-		memberService.validateAdminMemberByEmail(email);
+		MemberSummaryResponseDTO memberSummaryResponseDTO = memberService.getMemberSummary();
 
-		memberService.softDeleteByEmail(deleteEmail);
+		System.out.println(memberSummaryResponseDTO);
 
-		return ResponseEntity.status(HttpStatus.OK)
-			.body(new MessageResponseDTO(deleteEmail + SuccessMessage.DELETE.getMessage()));
-	}
+		model.addAttribute("memberSummary", memberSummaryResponseDTO);
 
-	@PatchMapping("/member/{email}/authority")
-	public ResponseEntity<?> changeAuthority(@PathVariable(name = "email") String email,
-		@RequestBody AuthorityRequestDTO authorityRequestDTO) {
+		return "/view/admin/summary";
 
-		Authority authority = authorityService.findByName(authorityRequestDTO.getAuthorityName());
-		Member member = memberService.findByEmail(email);
-
-		member.changeAuthority(authority);
-		memberService.save(member);
-
-		return ResponseEntity.status(HttpStatus.OK)
-			.body(new AuthorityResponseDTO(member.getEmail(), authority.getName()));
 	}
 
 }
