+++ src/main/java/kr/codeback/model/dto/response/MemberSummaryResponseDTO.java
@@ -1,0 +1,16 @@
+package kr.codeback.model.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@AllArgsConstructor
+@ToString
+public class MemberSummaryResponseDTO {
+
+	private long totalCount;
+	private long activeCount;
+	private long inactiveCount;
+
+}
